Refactor to use blueprints
